LPS-14447 Source formatting
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@68845 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portal-impl/src/com/liferay/portal/upgrade/BaseUpgradePortletPreferences.java
@@ -210,14 +210,14 @@ public abstract class BaseUpgradePortletPreferences extends UpgradeProcess {
 					Object[] group = getGroup(ownerId);
 
 					if (group != null) {
-						companyId = group[1];
+						companyId = (Long)group[1];
 					}
 				}
 				else {
 					Object[] layout = getLayout(plid);
 
 					if (layout != null) {
-						companyId = layout[1];
+						companyId = (Long)layout[1];
 					}
 				}
